requirements.txt
fastapi
uvicorn
gunicorn
azure-cosmos
